Multiple upgrades
Better loading time between the strating screen and the main screen.
Change the colour of the upgrade when you have enough points to buy it.
The songs are chosen randomly and not played again until all songs have been played. New button in the settings to play the next song.